+++ lib/aspera/schema.rb
@@ -7,14 +7,26 @@ require 'aspera/yaml'
 
 module Aspera
   # base class for plugins modules
-  module Schema
+  class Schema
     include Singleton
+
+    class << self
+      def known?(sym)
+        LOCATIONS.key?(sym)
+      end
+    end
 
     LOCATIONS = {
       spec: 'aspera/transfer/spec.schema.yaml',
       args: 'aspera/sync/args.schema.yaml',
-      conf: 'aspera/sync/conf.schema.yaml'
+      conf: 'aspera/sync/conf.schema.yaml',
+      opts: 'aspera/cli/options.schema.yaml'
     }
+
+    TRANSFER_INFO = 'opts:components.schemas.TransferInfo'
+    TRANSFER_SPEC = 'spec'
+    SYNC_CONF = 'conf'
+    SYNC_ARGS = 'args'
 
     def initialize
       @cache = {}
@@ -22,12 +34,16 @@ module Aspera
     end
 
     # Read schema from file or from cache
-    # @param sym [Symbol] :spec, :args, :conf
-    # @return [Hash] schema
-    def schema(sym)
-      return @cache[sym] if @cache.key?(sym)
-      Aspera.assert(LOCATIONS[sym])
-      @cache[sym] = Yaml.safe_load(File.join(@main_folder, LOCATIONS[sym]))
+    # @param path [String] <name>[:<path>]
+    # @return [Hash, nil] schema
+    def schema(path)
+      name, path = path.split(':', 2)
+      sym = name.to_sym
+      Aspera.assert(Schema.known?(sym)){"schema: #{sym}"}
+      @cache[sym] = Yaml.safe_load(File.read(File.join(@main_folder, LOCATIONS[sym]))) if !@cache.key?(sym)
+      schema = @cache[sym]
+      schema = schema.dig(*path.split('.')) unless path.nil? || path.empty?
+      schema
     end
   end
 end