+++ lib/aspera/transfer/spec.rb
@@ -50,7 +50,7 @@ module Aspera
           transfer_spec['resume_policy'] = POLICY_FIX[transfer_spec['resume_policy']] if transfer_spec.key?('resume_policy')
         end
       end
-      SCHEMA = CommandLineBuilder.read_schema(:spec, ascp: true)
+      SCHEMA = CommandLineBuilder.read_schema(Schema::TRANSFER_SPEC, ascp: true)
       # define constants for enums of parameters: <parameter>_<enum>, e.g. CIPHER_AES_128, DIRECTION_SEND, ...
       SCHEMA['properties'].each do |name, description|
         next unless description['enum'].is_a?(Array)