+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Released: [Place date of release here]
 ### New Features
 
 * `aoc`: New sub command in `shared_folder` : `admin workspace shared_folder <workspace ID> node <shared folder ID>`, shortcut to node operations directly from the shared folder.
+* **general**: Possibility to get Extended Value option schema using special value: `help`. e.g. : `--ts=help`.
 
 ### Issues Fixed